Fix: make Description component more reusable
- split Description in Section and ExpandableText components
- apply changes to places with old usage
