+++ packages/frontapp/src/app/components/DonateInfo/index.tsx
@@ -17,7 +17,8 @@ import AcceptRejectModals from '../AcceptRejectModals';
 import StyledButton from '../Button';
 import EmptyCover from '../../../assets/MOC-data/EmptyCover.png';
 import { Accepted, Rejected } from '../OneDonatedBookPreview';
-import DescriptionDetails from '../BookInfo/Description';
+import Section from '../Section';
+import ExpandableText from '../ExpandableText';
 
 const DonateInfo = ({
   description,
@@ -68,7 +69,9 @@ const DonateInfo = ({
           {statusInfo?.status === 'Free' && <Accepted>Accepted</Accepted>}
           {statusInfo?.status === 'Rejected' && <Rejected>Rejected</Rejected>}
         </ShortDescriptionWrapper>
-        <DescriptionDetails>{description}</DescriptionDetails>
+        <Section title={'Description: '}>
+          <ExpandableText>{description}</ExpandableText>
+        </Section>
       </BookHolder>
       <AcceptRejectModals
         active={active}
